Create member.txt
The member.txt file is a text file used to record the names of all group members. It is created by the group leader in the GitHub repository so that each member can add their full name.

This file helps to show that every group member has cloned the repository, made changes, committed with a message, and pushed their update to GitHub. It is used to track each member’s participation in the project using Git.